refactor statsig check
